Fix app installation when installLocation=preferExternal
The Nexus 5 does not have an SD card and the external storage device is
therefore emulated (/storage/emulated/). It is possible to use an USB-OTG cable
to add more storage (/storage/usbdisk).

Since commit 50bac1a, the package manager tries to install a package with
installLocation="preferExternal" to external storage. This fails with:

    E/Vold    (  189): Error creating imagefile (Read-only file system)
    E/Vold    (  189): ASEC image file creation failed (Read-only file system)
    W/Vold    (  189): Returning OperationFailed - no handler for errno 30
    E/PackageHelper( 4399): Failed to create secure container smdl728528075.tmp
    W/DefContainer( 4399): Failed to copy package at /data/local/tmp/biz.codefuture.svgviewer_7.apk
    W/DefContainer( 4399): java.io.IOException: Failed to create container smdl728528075.tmp
    W/DefContainer( 4399):       at com.android.defcontainer.DefaultContainerService.copyPackageToContainerInner(DefaultContainerService.java:327)
    W/DefContainer( 4399):       at com.android.defcontainer.DefaultContainerService.access$000(DefaultContainerService.java:67)
    W/DefContainer( 4399):       at com.android.defcontainer.DefaultContainerService$1.copyPackageToContainer(DefaultContainerService.java:108)
    W/DefContainer( 4399):       at com.android.internal.app.IMediaContainerService$Stub.onTransact(IMediaContainerService.java:60)
    W/DefContainer( 4399):       at android.os.Binder.execTransact(Binder.java:446)

This is caused by the wrong decision of PackageHelper.resolveInstallLocation()
to install to external storage. Because Environment.isNoEmulatedStorageExist()
returns true (/storage/usbdisk is found), it will get the size of the directory
returned by getSecondaryStorageDirectory().

Bug: this directory is /storage/emulated/0 which is definitely not the external
storage (is getSecondaryStorageDirectory() broken then?). As this location has
plenty of space, the function decides that external installation is possible...
while this is not the case.

The presented half-baked patch works around the issue by ignoring unmounted
devices (such as USB storage). When USB storage is attached however, the
installation fails again.

Either this whole "install to secondary storage" functionality should be
reverted, or more fixes needs to be done:

 - Ignore unmounted devices (this patch).
 - Check whether external media is mounted at /mnt/secure/asec before using it.
 - Make getSecondaryStorageDirectory() return no devices if there is no
   secondary device, or check whether it is not emulated (and mounted).
 - (other bugs?)

The bug was first introduced in cm-12.0:

    commit 8e46f42e6c35e9268a2bc7f82932794f8c841ddb
    Author: Xiaogang Cui <xiaogang@codeaurora.org>
    Date:   Mon Aug 18 11:27:10 2014 +0800

        Storage: Add secondary storage support for app installation

This was then subsequently fixed by:

    commit 5989b5eb07260cd1db2088418c5a2e048acc24a3
    Author: Thomas Wendt <thoemy@gmx.net>
    Date:   Wed Jan 14 16:18:19 2015 +0100

        Storage: Fix moving apps to external storage

but it broke again in cm-12.0 after:

    commit 50bac1a98d0e665252f5161d0e8c377d08bddbb6
    Author: Christopher R. Palmer <crpalmer@gmail.com>
    Date:   Wed Feb 18 09:12:33 2015 -0500

        Storage; Fix apps2sd for noemulated secondary storage

Tested using
https://f-droid.org/repository/browse/?fdid=biz.codefuture.svgviewer with and
without external USB storage (via USB-OTG).

Change-Id: I356feedeb41240da0120fc261784500b7111d081