TorchService: add a configurable overlay for config_TorchEnabled
Default is true. Devices like fugu that have a USB camera HAL that
enumerates the cameras based on build.prop settings can overide by
setting this to false. On these devices the USB OTG camera is only
available when it is plugged in.  This allows these devices to not
initialize the TorchService as there is never a torch compatible camera.

Change-Id: Ideea6931979b87725fdd321d5a2ab447aa1e231a

Conflicts:
	core/res/res/values/config.xml
	services/core/java/com/android/server/TorchService.java